show manage users within app share dialog
